src/light/color_light: increase cell capacity
This increases the number of allow animation cells from 256 to 65536.
256 was not enough to, for example, animate each hue in the HSV circle
moving 1 degree at a time (it would take 360 cells).

+++ lib/pbio/src/light/color_light.h
@@ -26,7 +26,7 @@ struct _pbio_color_light_t {
     const pbio_color_light_funcs_t *funcs;
     const void *cells;
     uint16_t interval;
-    uint8_t current_cell;
+    uint16_t current_cell;
 };
 
 void pbio_color_light_init(pbio_color_light_t *light, const pbio_color_light_funcs_t *funcs);
